// src/lib/types.ts
export type Rocket = {
    id: number
    image: string
    information: string
    name: string
    group_letter: string
    number: number
    group: string
    year_built: number
    weight: number
    max_earth_distance: number
    max_speed: number
    development_cost: number
    mission_duration: number
    payload: number
    involved_nations: number
    mission_goal: string
    lead_organization: string
    '': string
}